VSC Live Server vs Chrome Issue
VSCs' "Live Server" is formatting my page differently to what Chrome is presenting.
eg:
In my last upload VSC shows my social media links at the footer to be perfect size and padding. But when I open my page from GitHub the icons are enormous...
I am trying to play with the sizing and will have to upload a few more versions of code specifically just for the social media icons at the bottom to be the correct size AFTER they have been uploaded to GitHub.